Relinquish V2 components into older components
